app/requirements.txt
gunicorn
flask
pyyaml

